refactor(tour-map): remove polyline functionality and related code
The polyline functionality was removed as it is no longer required in the tour map component. This includes deleting the polyline component from the template and removing related code in the TypeScript file, such as the `polylinePath` array and the `createSmoothPath` method. The `addMarker` method was also refactored to exclude polyline-related logic.